Do not cache flow data as JSON by default
This functionality can become useful for debugging or otherwise,
but it's currently not used, so it makes sense to disable it.
Before this change flow data was store for L7 flows with tracing
headers, but that seems to be of even less potential usefulness,
so can be factored out completely.

Overall, this change should bring some performance improvements.