Rework linkage to traces parsed from headers
It turned out that linked spans cannot easily be extracted by the
functions that the SDK currently offers, e.g. `SpanContext` does
not even contain the ID of the parent span.
And the logic that was implemented up until now was doing it all
wrong and ended-up linking each trace ID that it parsed back to
itself (see #69).
With the need for smarter linking in mind (#70), it appears that
actually the right thing to do is to keep generating new traces
for HTTP flows, however link to existing spans whenever tracing
headers are present.